Refactoring de l'architecture des dossiers 'components' et migration de vue-touch-event via le système de plugins de nuxt.config
