+++ components/global/MagicTitle.vue
