+++ components/layout/scrollView/ScrollView.vue
@@ -14,10 +14,6 @@
 </template>
 
 <script>
-import Vue from "vue";
-import Vue2TouchEvents from "vue2-touch-events";
-Vue.use(Vue2TouchEvents);
-
 export default {
   data() {
     return {
