+++ plugins/vue-touch-events.js
@@ -1,0 +1,4 @@
+import Vue from "vue";
+import Vue2TouchEvents from "vue2-touch-events";
+
+Vue.use(Vue2TouchEvents);
